Extract an "instqueue" program from instpackage.
Packagers need a way to make the queue (and only the queue) at
package-install time.
